requirements.txt
# IBKR API
ib_insync==0.9.86

# Data handling
pandas==2.1.4
numpy==1.26.2

# Web dashboard
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Scheduling
schedule==1.2.0
python-dateutil==2.8.2

# Database
sqlalchemy==2.0.23

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
colorlog==6.8.0

# Utilities
pytz==2023.3
